requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
amqp==2.6.1
    # via kombu
attrs==19.3.0
    # via
    #   -r requirements.in
    #   jsonschema
backoff==1.10.0
    # via -r requirements.in
backports-abc==0.5
    # via -r requirements.in
billiard==3.6.4.0
    # via celery
boto3==1.17.75
    # via django-s3-storage
botocore==1.20.75
    # via
    #   boto3
    #   s3transfer
cachetools==4.2.2
    # via google-auth
celery==4.4.7
    # via -r requirements.in
certifi==2020.12.5
    # via
    #   minio
    #   requests
    #   sentry-sdk
cffi==1.14.5
    # via google-crc32c
chardet==4.0.0
    # via requests
confusable-homoglyphs==3.2.0
    # via django-registration
django-bulk-update==2.2.0
    # via -r requirements.in
django-cte==1.1.5
    # via -r requirements.in
django-db-readonly==0.7.0
    # via -r requirements.in
django-filter==2.4.0
    # via -r requirements.in
django-js-asset==1.2.2
    # via django-mptt
django-js-reverse==0.9.1
    # via -r requirements.in
django-mathfilters==1.0.0
    # via -r requirements.in
django-model-utils==4.0.0
    # via -r requirements.in
django-mptt==0.11.0
    # via -r requirements.in
django-postmark==0.1.6
    # via -r requirements.in
django-prometheus==2.1.0
    # via -r requirements.in
django-redis==4.12.1
    # via -r requirements.in
django-registration==3.1.1
    # via -r requirements.in
django-s3-storage==0.13.4
    # via -r requirements.in
django-webpack-loader==0.7.0
    # via -r requirements.in
django==2.2.17
    # via
    #   -r requirements.in
    #   django-bulk-update
    #   django-db-readonly
    #   django-filter
    #   django-js-reverse
    #   django-model-utils
    #   django-mptt
    #   django-redis
    #   django-registration
    #   django-s3-storage
    #   djangorestframework
    #   jsonfield
djangorestframework==3.12.1
    # via -r requirements.in
future==0.18.2
    # via -r requirements.in
google-api-core[grpc]==1.27.0
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-error-reporting
    #   google-cloud-kms
    #   google-cloud-logging
google-api-python-client==2.4.0
    # via -r requirements.in
google-auth-httplib2==0.1.0
    # via google-api-python-client
google-auth-oauthlib==0.4.4
    # via gspread
google-auth==1.30.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
    #   google-cloud-core
    #   google-cloud-storage
    #   gspread
google-cloud-core==1.6.0
    # via
    #   -r requirements.in
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-error-reporting==1.1.2
    # via -r requirements.in
google-cloud-kms==0.2.1
    # via -r requirements.in
google-cloud-logging==2.3.1
    # via google-cloud-error-reporting
google-cloud-storage==1.38.0
    # via -r requirements.in
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==1.3.0
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.53.0
    # via
    #   google-api-core
    #   grpc-google-iam-v1
grpc-google-iam-v1==0.11.4
    # via google-cloud-kms
grpcio==1.37.1
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
gspread==3.6.0
    # via -r requirements.in
gunicorn==19.6.0
    # via -r requirements.in
html5lib==1.1
    # via -r requirements.in
httplib2==0.19.1
    # via
    #   django-postmark
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
idna==2.10
    # via requests
importlib-metadata==1.7.0
    # via
    #   -r requirements.in
    #   jsonschema
    #   kombu
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonfield==2.0.2
    # via -r requirements.in
jsonschema==3.2.0
    # via -r requirements.in
kombu==4.6.11
    # via celery
le-utils==0.1.30
    # via -r requirements.in
minio==3.0.3
    # via -r requirements.in
newrelic==6.2.0.156
    # via -r requirements.in
oauth2client==4.1.3
    # via -r requirements.in
oauthlib==3.1.0
    # via requests-oauthlib
packaging==20.9
    # via google-api-core
pathlib==1.0.1
    # via -r requirements.in
pillow==8.1.1
    # via -r requirements.in
progressbar2==3.38.0
    # via -r requirements.in
prometheus-client==0.10.1
    # via django-prometheus
proto-plus==1.18.1
    # via
    #   google-cloud-error-reporting
    #   google-cloud-logging
protobuf==3.17.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psycopg2-binary==2.8.6
    # via -r requirements.in
pyasn1-modules==0.2.8
    # via
    #   google-auth
    #   oauth2client
pyasn1==0.4.8
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pycountry==17.5.14
    # via
    #   -r requirements.in
    #   le-utils
pycparser==2.20
    # via cffi
pyparsing==2.4.7
    # via
    #   httplib2
    #   packaging
pyrsistent==0.17.3
    # via jsonschema
python-dateutil==2.8.1
    # via
    #   -r requirements.in
    #   botocore
python-postmark==0.5.0
    # via -r requirements.in
python-utils==2.5.6
    # via progressbar2
pytz==2021.1
    # via
    #   celery
    #   django
    #   django-postmark
    #   google-api-core
    #   minio
raven==6.10.0
    # via -r requirements.in
redis==3.5.3
    # via
    #   -r requirements.in
    #   django-redis
requests-oauthlib==1.3.0
    # via google-auth-oauthlib
requests==2.25.1
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-storage
    #   gspread
    #   requests-oauthlib
rsa==4.7.2
    # via
    #   google-auth
    #   oauth2client
s3transfer==0.4.2
    # via boto3
sentry-sdk==1.1.0
    # via -r requirements.in
six==1.16.0
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-resumable-media
    #   grpcio
    #   html5lib
    #   jsonschema
    #   oauth2client
    #   progressbar2
    #   protobuf
    #   python-dateutil
    #   python-utils
sqlparse==0.4.1
    # via django
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.4
    # via
    #   botocore
    #   minio
    #   requests
    #   sentry-sdk
vine==1.3.0
    # via
    #   amqp
    #   celery
webencodings==0.5.1
    # via html5lib
whitenoise==4.1.4
    # via -r requirements.in
zipp==3.4.1
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
